Add Tailwind CSS and restructure components
Integrated Tailwind CSS by adding configuration files and updating dependencies. Refactored the app structure by creating new components (Header, Hero, About, Skills, Projects, Contact), updated App.jsx to use them, and removed the old Practice file. Also updated the HTML title and added a backup of the previous App.jsx as App_old.jsx.
