+++ package.json
@@ -26,7 +26,7 @@
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^16.4.0",
     "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.13",
+    "tailwindcss": "^3.4.0",
     "vite": "npm:rolldown-vite@7.1.12"
   },
   "overrides": {
