+++ tailwind.config.js
@@ -1,0 +1,82 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
+  theme: {
+    extend: {
+      colors: {
+        gray: {
+          50: '#f9fafb',
+          100: '#f3f4f6',
+          200: '#e5e7eb',
+          300: '#d1d5db',
+          400: '#9ca3af',
+          500: '#6b7280',
+          600: '#4b5563',
+          700: '#374151',
+          800: '#1f2937',
+          900: '#111827',
+        },
+        blue: {
+          50: '#eff6ff',
+          100: '#dbeafe',
+          200: '#bfdbfe',
+          400: '#60a5fa',
+          500: '#3b82f6',
+          600: '#2563eb',
+          700: '#1d4ed8',
+        },
+        purple: {
+          100: '#f3e8ff',
+          500: '#a855f7',
+          600: '#9333ea',
+        },
+        yellow: {
+          400: '#facc15',
+        },
+        orange: {
+          500: '#f97316',
+        },
+        indigo: {
+          600: '#4f46e5',
+        },
+        green: {
+          400: '#4ade80',
+          500: '#22c55e',
+        },
+        emerald: {
+          600: '#059669',
+        },
+        red: {
+          500: '#ef4444',
+        },
+        pink: {
+          600: '#db2777',
+        },
+        gray: {
+          100: '#f3f4f6',
+          200: '#e5e7eb',
+          300: '#d1d5db',
+          400: '#9ca3af',
+          500: '#6b7280',
+          600: '#4b5563',
+          700: '#374151',
+          800: '#1f2937',
+          900: '#111827',
+        }
+      },
+      fontFamily: {
+        'sans': ['Inter', 'system-ui', 'sans-serif'],
+        'mono': ['Fira Code', 'Monaco', 'Cascadia Code', 'Roboto Mono', 'monospace'],
+      },
+      animation: {
+        'float': 'float 3s ease-in-out infinite',
+        'glow': 'glow 2s ease-in-out infinite alternate',
+        'bounce-slow': 'bounce 3s infinite',
+      },
+    },
+  },
+  plugins: [],
+}
